Tag branch build revisions as multiarch images on dockerhub
This adds a tag prefix "devel-" and adds a buildkite annotation with
the resulting tag.
